move FourVectorHLT.h into the plugins/ directory
