--- DQM/HLTEvF/BuildFile.xml
@@ -1,20 +1,0 @@
-<use   name="DQMServices/Core"/>
-<use   name="DataFormats/Common"/>
-<use   name="DataFormats/HLTReco"/>
-<use   name="DataFormats/Provenance"/>
-<use   name="DataFormats/BTauReco"/>
-<use   name="DataFormats/TauReco"/>
-<use   name="FWCore/Framework"/>
-<use   name="FWCore/MessageLogger"/>
-<use   name="FWCore/ParameterSet"/>
-<use   name="FWCore/ServiceRegistry"/>
-<use   name="FWCore/Services"/>
-<use   name="FWCore/Utilities"/>
-<use   name="Geometry/CaloGeometry"/>
-<use   name="Geometry/CaloTopology"/>
-<use   name="RecoEcal/EgammaCoreTools"/>
-<use   name="HLTrigger/HLTcore"/>
-<use   name="rootcore"/>
-<export>
-  <lib   name="1"/>
-</export>
