+++ DQM/HLTEvF/plugins/FourVectorHLT.cc
@@ -1,16 +1,14 @@
-// See header file for information. 
 #include "FWCore/Framework/interface/EDAnalyzer.h"
-#include "DataFormats/Common/interface/Handle.h"
 #include "FWCore/Framework/interface/Run.h"
 #include "FWCore/Framework/interface/MakerMacros.h"
-
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
-#include "DQM/HLTEvF/interface/FourVectorHLT.h"
-
+#include "DataFormats/Common/interface/Handle.h"
 #include "DataFormats/HLTReco/interface/TriggerObject.h"
 #include "DataFormats/HLTReco/interface/TriggerEvent.h"
-
 #include "DQMServices/Core/interface/MonitorElement.h"
+
+// see header file for information. 
+#include "FourVectorHLT.h"
 
 using namespace edm;
